Check for Existing User: Queries the database to check if a user with the same username already exists. If an existing user is found, it returns a 400 status with a message indicating that the user already exists.
Save User to Database: Saves the new user to the database using newUser.save().

Generate JWT Token: Generates a JWT token containing the user's username and id using jsonwebtoken.sign(). The token is signed with a secret key (process.env.JWTKEY) and has an expiration time of 1 hour ({ expiresIn: "1h" }).

Return Response: Returns a 200 status along with the user object and the generated JWT token in the response body.