Find User in Database: Queries the database to find a user with the provided username using UserModel.findOne().
Check Password Validity: Compares the provided password with the hashed password stored in the database using bcrypt.compare(). If the passwords match, it proceeds; otherwise, it returns a 400 status with a message indicating that the password is incorrect.

Generate JWT Token: Generates a JWT token similar to the register user function.

Return Response: Returns a 200 status along with the user object and the generated JWT token in the response body if the login is successful. If the user is not found, it returns a 404 status with a message indicating that the user was not found.