Retrieve Token from Request Header: Extracts the JWT token from the Authorization header of the incoming request using req.headers.authorization. The token is typically sent in the format Bearer <token>.
Verify JWT Token: Verifies the authenticity of the JWT token by calling jwt.verify() method. It uses the token and the secret key (secret) retrieved from the environment variables. If the token is valid, the method returns the decoded payload of the token.

Attach User ID to Request Body: If a valid token is present, extracts the user ID (id) from the decoded payload and attaches it to the request body (req.body._id). This allows subsequent middleware functions or route handlers to access the user ID for further processing.

Call Next Middleware Function: Calls the next() function to pass control to the next middleware function in the request-response cycle. This is necessary to proceed with the execution of subsequent middleware functions or route handlers.