Field Definitions:
username: Specifies a field for the user's username. It is of type String and is required.
password: Specifies a field for the user's password. It is of type String and is required.
firstname: Specifies a field for the user's first name. It is of type String and is required.
lastname: Specifies a field for the user's last name. It is of type String and is required.
isAdmin: Specifies a field for indicating whether the user is an administrator. It is of type Boolean and has a default value of false.
profilePicture: Specifies a field for the URL of the user's profile picture. It is of type String.
coverPicture: Specifies a field for the URL of the user's cover picture. It is of type String.
about: Specifies a field for a short bio or description about the user. It is of type String.
livesIn: Specifies a field for the user's location or where they live. It is of type String.
worksAt: Specifies a field for the user's workplace or company. It is of type String.
relationship: Specifies a field for the user's relationship status. It is of type String.
country: Specifies a field for the user's country. It is of type String.
followers: Specifies a field for an array containing the IDs of users who follow this user.
following: Specifies a field for an array containing the IDs of users whom this user follows.