Get User Function: getUser Function: Retrieves a user's data by their ID from the MongoDB collection using UserModel.findById().
Response Handling: If the user is found, it removes the password field from the user document before sending the other details in the response. If the user is not found, it returns a 404 status with a message indicating that the user was not found. If an error occurs during the database query, it returns a 500 status with the error message.