Field Definitions:
userId: Specifies a field for the ID of the user who created the post. It is of type String and is required.
desc: Specifies a field for the description or content of the post. It is of type String and is required.
likes: Specifies a field for an array containing the IDs of users who liked the post.
createdAt: Specifies a field for the creation timestamp of the post. It is of type Date and defaults to the current date and time when a new post is created.
image: Specifies a field for the URL of an image attached to the post. It is of type String.