Route Definitions: Create Post Route: Defines a POST route for creating a new post. When a POST request is made to /, the createPost controller function is invoked to handle the request.
Get Post by ID Route: Defines a GET route for retrieving a post by its ID. When a GET request is made to /:id, the getPost controller function is invoked to handle the request.

Update Post Route: Defines a PUT route for updating a post by its ID. When a PUT request is made to /:id, the updatePost controller function is invoked to handle the request.

Delete Post Route: Defines a DELETE route for deleting a post by its ID. When a DELETE request is made to /:id, the deletePost controller function is invoked to handle the request.

Like/Dislike Post Route: Defines a PUT route for liking or disliking a post by its ID. When a PUT request is made to /:id/like, the likePost controller function is invoked to handle the request.

Get Timeline Posts Route: Defines a GET route for retrieving timeline posts for a user by their ID. When a GET request is made to /:id/timeline, the getTimelinePosts controller function is invoked to handle the request.