Route Definitions: Create Chat Route: Defines a POST route for creating a new chat. When a POST request is made to /, the createChat controller function is invoked to handle the request.
User Chats Route: Defines a GET route for retrieving chats for a specific user. When a GET request is made to /:userId, the userChats controller function is invoked to handle the request.

Find Chat Route: Defines a GET route for finding a chat between two specific users. When a GET request is made to /find/:firstId/:secondId, the findChat controller function is invoked to handle the request.