MessageSchema Definition: The new mongoose.Schema() function defines a schema for the message model. It specifies the structure of documents in the "Message" collection in MongoDB.
chatId Field: It defines a field named chatId which is of type String. This field stores the ID of the chat to which the message belongs.

senderId Field: It defines a field named senderId which is of type String. This field stores the ID of the user who sent the message.

text Field: It defines a field named text which is of type String. This field stores the text content of the message.