Defining API Functions logIn Function
export const logIn = (formData) => API.post('/auth/login', formData);
This function sends a POST request to the endpoint /auth/login with the provided formData.
formData is an object containing the login details (usually a username/email and password) that the user has input.
The API.post method is used to send the request. Since the Axios instance API has a base URL set, the full URL for this request will be http://localhost:5000/auth/login.