Defining API Functions createChat Function
export const createChat = (data) => API.post('/chat/', data);
This function sends a POST request to the endpoint /chat/ with the provided data.
data is an object containing the information needed to create a new chat (such as participant IDs, initial messages, etc.).
The API.post method is used to send the request. Since the Axios instance API has a base URL set, the full URL for this request will be http://localhost:5000/chat/.