userChats Function
export const userChats = (id) => API.get(/chat/${id});
This function sends a GET request to the endpoint /chat/{id} to retrieve all chat data for a specific user.
id is the user ID whose chats you want to retrieve.
The API.get method is used to send the request, and the full URL for this request will be http://localhost:5000/chat/{id}.