Defining API Functions getMessages Function
export const getMessages = (id) => API.get(/message/${id});
This function sends a GET request to the endpoint /message/{id} to retrieve all messages for a specific chat or conversation.
id is the identifier of the chat or conversation for which you want to retrieve messages.
The API.get method is used to send the request. Since the Axios instance API has a base URL set, the full URL for this request will be http://localhost:5000/message/{id}.