addMessage Function
export const addMessage = (data) => API.post('/message/', data);
This function sends a POST request to the endpoint /message/ to add a new message to a chat or conversation.
data is an object containing the information needed to create a new message (such as the message content, sender ID, chat ID, timestamp, etc.).
The API.post method is used to send the request, and the full URL for this request will be http://localhost:5000/message/.