likePost Function
export const likePost = (id, userId) => API.put(posts/${id}/like, { userId: userId });
This function sends a PUT request to the endpoint /posts/{id}/like to like a specific post.
id is the identifier of the post to be liked.
userId is the identifier of the user who is liking the post.
The request payload contains an object with the userId. The API.put method is used to send the request, and the full URL for this request will be http://localhost:5000/posts/{id}/like.