uploadPost Function
export const uploadPost = (data) => API.post("/posts", data);
This function sends a POST request to the endpoint /posts to upload a new post.
data is an object containing the information needed to create a new post (such as the post content, images, etc.).
The API.post method is used to send the request, and the full URL for this request will be http://localhost:5000/posts.