API Functions getUser Function
export const getUser = (userId) => API.get(/user/${userId});
This function sends a GET request to the endpoint /user/{userId} to retrieve the details of a specific user.
userId is the ID of the user whose details are being fetched.
The API.get method is used to send the request. Since the Axios instance API has a base URL set, the full URL for this request will be http://localhost:5000/user/{userId}.