updateUser Function
export const updateUser = (id, data) => API.put(/user/${id}, data);
This function sends a PUT request to the endpoint /user/{id} to update the details of a specific user.
id is the ID of the user whose details are being updated.
data is the payload for the PUT request, containing the updated user details.
The API.put method is used to send the request. Since the Axios instance API has a base URL set, the full URL for this request will be http://localhost:5000/user/{id}.