getAllUser Function
export const getAllUser = () => API.get('/user');
This function sends a GET request to the endpoint /user to retrieve the details of all users.
The API.get method is used to send the request. Since the Axios instance API has a base URL set, the full URL for this request will be http://localhost:5000/user.