followUser Function
export const followUser = (id, data) => API.put(/user/${id}/follow, data);
This function sends a PUT request to the endpoint /user/{id}/follow to follow a specific user.
id is the ID of the user to be followed.
data is the payload for the PUT request, possibly containing the details of the user who is following.
The API.put method is used to send the request. Since the Axios instance API has a base URL set, the full URL for this request will be http://localhost:5000/user/{id}/follow.