unfollowUser Function
export const unfollowUser = (id, data) => API.put(/user/${id}/unfollow, data);
This function sends a PUT request to the endpoint /user/{id}/unfollow to unfollow a specific user.
id is the ID of the user to be unfollowed.
data is the payload for the PUT request, possibly containing the details of the user who is unfollowing.
The API.put method is used to send the request. Since the Axios instance API has a base URL set, the full URL for this request will be http://localhost:5000/user/{id}/unfollow.
Summary
Axios Library: Used for making HTTP requests.
Axios Instance: API is an Axios instance configured with a base URL, making it easier to manage and reuse the configuration.
Request Interceptor: Adds an Authorization header to each request if a token is available in localStorage.
This ensures that all requests made with this instance will include the user's token, which is often required for authentication and authorization purposes.
API Functions:
getUser: Retrieves details of a specific user.
updateUser: Updates details of a specific user.
getAllUser: Retrieves details of all users.
followUser: Follows a specific user.
unfollowUser: Unfollows a specific user.
By organizing the code in this way, you ensure that all requests are consistently configured, and you have simple, reusable functions for interacting with the user-related endpoints that automatically include the necessary authentication token if available.