Async Thunks:
Async thunks are functions that can contain asynchronous logic and dispatch actions based on the result.
Three async thunks are defined:
createChat: For creating a new chat.
userChats: For retrieving chats associated with a user.
findChat: For finding a chat between two users.
Each async thunk is created using createAsyncThunk, which takes a name for the action and an async function.
Inside each async function, API calls are made using functions from ChatApi. These functions are awaited, and their responses are returned.
In case of an error, rejectWithValue is used to return the error payload.