Async Thunks:
Async thunks are functions that can contain asynchronous logic and dispatch actions based on the result.
Two async thunks are defined:
getMessages: For retrieving messages.
addMessage: For adding a new message.
Each async thunk is created using createAsyncThunk, which takes a name for the action and an async function.
Inside each async function, API calls are made using functions from MessageApi. These functions are awaited, and their responses are returned.
In case of an error, rejectWithValue is used to return the error payload.