Slice:
A slice is a collection of reducer logic and actions for a specific slice of the Redux state.
The messageSlice is created using createSlice, which takes an object with configuration options.
It defines the name of the slice (name) and the initial state (initialState).
The initial state includes an empty array messages, loading set to false, and error set to null.
Reducer functions can be added directly to the reducers field of the createSlice call. In this case, no reducer functions are defined here.
Instead, the extraReducers field is used to define reducer logic for actions dispatched by the async thunks.
Each addCase method attaches reducer logic to a specific action type:
For each async thunk, there are three cases: pending, fulfilled, and rejected.
The pending case sets loading to true and clears any previous errors.
The fulfilled case updates the state with the retrieved data or clears the loading state if it's for adding a message.
The rejected case sets loading to false and updates the error state with the error payload.