Async Thunk:
An async thunk named uploadImage is defined.
It is created using createAsyncThunk, taking a name for the action and an async function.
Inside the async function, an API call is made using UploadApi.uploadImage(data).
If the API call is successful, nothing specific happens, but if it fails, the rejectWithValue function is used to return the error payload.