Slice:
A slice named userSlice is created using createSlice.
It defines the name of the slice (name) and the initial state (initialState).
The initial state includes userData set to null, loading set to false, and error set to null.
No reducer functions are defined in the reducers field of the createSlice call.
Instead, the extraReducers field is used to define reducer logic for actions dispatched by the async thunks.
For each async thunk, reducer cases are added to handle the pending, fulfilled, and rejected states.
The pending case sets loading to true and clears any previous errors.
The fulfilled case updates the state accordingly based on the action payload.
The rejected case sets loading to false and updates the error state with the error payload.