ReactDOM.render:
ReactDOM.render() is used to render the application to the DOM.
The Provider component wraps the entire application, providing access to the Redux store.
Inside the Provider, the BrowserRouter component is used to provide routing capabilities to the application.
Within the BrowserRouter, the Routes component is used to define the routes of the application.
The Route component specifies a wildcard path "*" that matches any route and renders the App component for all routes.
The rendered content is appended to the DOM element with the ID "root".