The CSS styles provided define the appearance and layout of the InfoCard component. Let's break down each part:
.InfoCard:

display: flex;: Sets the container to a flex layout, allowing its children to be laid out in a column.
flex-direction: column;: Specifies that the children elements should be stacked vertically.
gap: 0.75rem;: Adds a small gap between each child element.
background-color: var(--cardColor);: Sets the background color of the card using a CSS variable.
padding: 1rem;: Adds padding around the content of the card.
border-radius: 1rem;: Rounds the corners of the card, giving it a softer appearance.
width: 90%;: Sets the width of the card to take up 90% of its containing element's width.
.infoHead:

display: flex;: Configures the container for the header section to use flex layout.
justify-content: space-between;: Positions the children elements with space between them horizontally, pushing them to the opposite edges of the container.
align-items: center;: Aligns the children elements vertically in the center of the container.
.infoHead > div
:

cursor: pointer;: Changes the cursor to a pointer when hovering over the child div elements within the .infoHead section, indicating interactivity.
.logout-button:

width: 7rem;: Sets the width of the log out button to 7rem.
height: 2rem;: Sets the height of the log out button to 2rem.
margin-top: 6rem;: Adds top margin to the log out button to create space between the profile information and the button.
align-self: flex-end;: Aligns the log out button to the end of the flex container vertically.
These styles collectively create a visually appealing and well-structured profile information card, with appropriate spacing, alignment, and interactive elements.